fastapi/requirements.txt
fastapi
uvicorn
redis
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-jaeger
opentelemetry-instrumentation-fastapi
